backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pyreadstat>=1.2.7",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[tool.uv]
package = false
